+++ src/utils/onboard.ts
@@ -1,143 +1,126 @@
-// import React from "react";
-import Onboard from "bnc-onboard";
-import { Wallet, Initialization, Ens } from "bnc-onboard/dist/src/interfaces";
-import { ethers } from "ethers";
-import {
-  onboardApiKey,
-  hubPoolChainId,
-  ChainId,
-  providerUrlsTable,
-  trackEvent,
-  debug,
-} from "utils";
-import { update, disconnect, error } from "state/connection";
-import { store } from "state";
+import { init } from "@web3-onboard/react";
+import injectedModule from "@web3-onboard/injected-wallets";
+import walletConnectModule from "@web3-onboard/walletconnect";
+import gnosisModule from "@web3-onboard/gnosis";
+import coinbaseModule from "@web3-onboard/coinbase";
+import { onboardApiKey, ChainId, providerUrlsTable } from "utils";
+import logo from "assets/across-logo-v2.svg";
 
-/* Onboard config */
-export function onboardBaseConfig(): Initialization {
-  return {
-    dappId: onboardApiKey,
-    networkId: hubPoolChainId,
-    hideBranding: true,
-    walletSelect: {
-      wallets: [
-        { walletName: "metamask", preferred: true },
-        {
-          walletName: "walletConnect",
-          rpc: {
-            [ChainId.MAINNET]: providerUrlsTable[ChainId.MAINNET],
-            [ChainId.OPTIMISM]: providerUrlsTable[ChainId.OPTIMISM],
-            [ChainId.BOBA]: providerUrlsTable[ChainId.BOBA],
-            [ChainId.ARBITRUM]: providerUrlsTable[ChainId.ARBITRUM],
-            [hubPoolChainId]: providerUrlsTable[hubPoolChainId],
-          },
-          preferred: true,
-        },
-        // { walletName: "coinbase", preferred: true },
-        { walletName: "tally", preferred: true },
-        { walletName: "detectedwallet" },
+const injected = injectedModule();
+const gnosis = gnosisModule();
+const walletConnect = walletConnectModule();
+const coinbase = coinbaseModule();
+
+/* 
+export interface InitOptions {
+    /**
+     * Wallet modules to be initialized and added to wallet selection modal
+     */
+//  wallets: WalletInit[];
+/**
+ * The chains that your app works with
+ */
+//  chains: Chain[] | ChainWithDecimalId[];
+/**
+ * Additional metadata about your app to be displayed in the Onboard UI
+ */
+//  appMetadata?: AppMetadata;
+/**
+ * Define custom copy for the 'en' locale or add locales to i18n your app
+ */
+//  i18n?: i18nOptions;
+/**
+ * Customize the connect modal
+ */
+//  connect?: ConnectModalOptions;
+/**
+ * Customize the account center UI
+ */
+//  accountCenter?: AccountCenterOptions;
+/**
+ * Opt in to Blocknative value add services (transaction updates) by providing
+ * your Blocknative API key, head to https://explorer.blocknative.com/account
+ */
+//  apiKey?: string;
+/**
+ * Transaction notification options
+ */
+//  notify?: Partial<NotifyOptions> | Partial<Notify>;
+/**Gas module */
+//  gas?: typeof gas;
+//  }
+
+/* 
+    export interface Chain {
+      namespace?: 'evm';
+      id: ChainId;
+      rpcUrl: string;
+      label: string;
+      token: TokenSymbol;
+      color?: string;
+      icon?: string;
+      providerConnectionInfo?: ConnectionInfo;
+      publicRpcUrl?: string;
+      blockExplorerUrl?: string;
+  }
+    */
+
+export function onboardInit() {
+  return init({
+    apiKey: onboardApiKey,
+    wallets: [injected, coinbase, walletConnect, gnosis],
+
+    chains: [
+      {
+        id: 1,
+        token: "ETH",
+        label: "Ethereum Mainnet",
+        rpcUrl: providerUrlsTable[ChainId.MAINNET],
+      },
+      {
+        id: 10,
+        token: "OP",
+        label: "Optimism Mainnet",
+        rpcUrl: providerUrlsTable[ChainId.OPTIMISM],
+      },
+      {
+        id: 137,
+        token: "MATIC",
+        label: "Polygon Mainnet",
+        rpcUrl: providerUrlsTable[ChainId.POLYGON],
+      },
+      {
+        id: 288,
+        token: "BOBA",
+        label: "BOBA Mainnet",
+        rpcUrl: providerUrlsTable[ChainId.BOBA],
+      },
+      {
+        id: 42161,
+        token: "ETH",
+        label: "Arbitrum Mainnet",
+        rpcUrl: providerUrlsTable[ChainId.ARBITRUM],
+      },
+    ],
+    appMetadata: {
+      name: "Across Bridge",
+      icon: logo,
+      description:
+        "Across is the fastest, cheapest and most secure cross-chain bridge for Ethereum, Arbitrum, Optimism, Polygon, Boba and other Layer 1 and Layer 2 networks. Transfer tokens with Across.",
+      recommendedInjectedWallets: [
+        { name: "Metamask", url: "https://metamask.io" },
+        { name: "Coinbase", url: "https://wallet.coinbase.com/" },
+        { name: "WalletConnect", url: "https://walletconnect.org/" },
+        { name: "Gnosis Safe", url: "https://gnosis.safe/" },
       ],
     },
-    walletCheck: [{ checkName: "connect" }, { checkName: "accounts" }],
-    // To prevent providers from requesting block numbers every 4 seconds (see https://github.com/WalletConnect/walletconnect-monorepo/issues/357)
-    blockPollingInterval: 1000 * 60 * 60,
-  };
-}
-export type Emit = (event: string, data?: any) => void;
-export function OnboardEthers(config: Initialization, emit: Emit) {
-  let savedWallet: Wallet | undefined;
-  const onboard = Onboard({
-    ...config,
-    subscriptions: {
-      address: (address: string) => {
-        emit("update", { account: address });
-        if (!address) {
-          // if we dont call reset here when account is undefined, we wont be able to reconnect
-          // this can happen if user locks their wallet.
-          reset();
-        }
+    accountCenter: {
+      desktop: {
+        enabled: false,
       },
-      network: (chainIdInHex) => {
-        if (chainIdInHex == null) {
-          return;
-        }
-        const chainId = ethers.BigNumber.from(chainIdInHex).toNumber();
-        // need to make new provider on chain change
-        if (savedWallet?.provider) {
-          // when chain change will always follow first wallet connect or chain change, so we can emit
-          // signer and provider here, so it only happens once.
-          const provider = new ethers.providers.Web3Provider(
-            savedWallet.provider
-          );
-          const signer = provider.getSigner();
-          emit("update", {
-            chainId,
-            provider,
-            signer,
-          });
-        } else {
-          emit("update", {
-            chainId,
-          });
-        }
-      },
-      wallet: (wallet: Wallet) => {
-        savedWallet = wallet;
-        if (wallet.provider) {
-          emit("update", {
-            account: wallet.provider.selectedAddress,
-          });
-        }
-      },
-      ens: (ens: Ens) => {
-        const ensName = ens?.name;
-        if (savedWallet?.provider) {
-          emit("update", {
-            ensName,
-          });
-        }
+      mobile: {
+        enabled: false,
       },
     },
   });
-  async function init() {
-    try {
-      await onboard.walletSelect();
-      await onboard.walletCheck();
-      emit("init");
-    } catch (err: any) {
-      emit(
-        "error",
-        new Error(("Could not initialize Onboard: " + err.message) as string)
-      );
-    }
-  }
-  async function reset() {
-    try {
-      await onboard.walletReset();
-      emit("disconnect");
-    } catch (err) {
-      emit("error", err);
-    }
-  }
-  return {
-    init,
-    reset,
-    onboard,
-  };
 }
-export const onboard = OnboardEthers(onboardBaseConfig(), (event, data) => {
-  if (debug) console.log("onboard event", event, data);
-  if (event === "init") {
-    trackEvent({ category: "wallet", action: "connect", name: "null" });
-  }
-  if (event === "update") {
-    store.dispatch(update(data));
-  }
-  if (event === "disconnect") {
-    store.dispatch(disconnect());
-    trackEvent({ category: "wallet", action: "disconnect", name: "null" });
-  }
-  if (event === "error") {
-    store.dispatch(error(data));
-  }
-});
