+++ src/views/Transactions/hooks/useMyTransactionsView.tsx
@@ -1,15 +1,14 @@
 import { useConnection } from "state/hooks";
 import useWindowSize from "hooks/useWindowsSize";
-import { onboard, BREAKPOINTS } from "utils";
+import { BREAKPOINTS } from "utils";
 
 import { useTxClient } from "./useTxClient";
 import { usePagination } from "./usePagination";
 
 export function useMyTransactionsView() {
-  const { init } = onboard;
   const { initialLoading, filledTransferTuples, pendingTransferTuples } =
     useTxClient();
-  const { isConnected, account } = useConnection();
+  const { isConnected, account, connect } = useConnection();
 
   const { width = 0 } = useWindowSize();
   const isMobile = width < BREAKPOINTS.laptopMin;
@@ -17,7 +16,7 @@ export function useMyTransactionsView() {
   const filledTransfersPagination = usePagination(filledTransferTuples.length);
 
   return {
-    connectWallet: init,
+    connectWallet: connect,
     account,
     initialLoading,
     pendingTransferTuples,
